requirements.txt
chalice
line-bot-sdk
pg8000
requests